# To load single DPAS B matrix instead of two per 2D block io instruction from the transposed memory 