… include/triton/Tools/Sys/GetEnv.hpp
… third_party/intel/lib/TritonIntelGPUToLLVM/LoadStoreOpToLLVM.cpp