To load single DPAS B matrix per 2D block io instruction from the column major matrix in memory gets better performance for flash attention.

Because unlike the row major matrix, the values, which includes more than one DPAS B operands returned by a single 2D transposed block IO,  cannot be used as DPAS operands directly.

We have to shuffle the value in the register before pass it to the DPAS instruction and this is not optimized by the IGC for now.