@alexbaden 2024-11-05
I am seeing a ~5% regression in `AxBT` performance and a ~2.5% regression in `ATxBT` performance with this change. Is that expected?

Before:
Geomean |   |  
-- | -- | --
Torch (ms) | Triton (ms) | Triton % Torch
0.316446402 | 0.393551599 | 80.41%
0.338533627 | 0.66412835 | 50.97%
0.318394668 | 1.600593032 | 19.89%
0.456415746 | 1.986145424 | 22.98%


After:
Torch (ms) | Triton (ms) | Triton % Torch
-- | -- | --
0.314741015 | 0.388807784 | 80.95%
0.337140929 | 0.751260794 | 44.88%
0.313591856 | 1.608697493 | 19.49%
0.451715958 | 2.195144218 | 20.58%