@etiotto 2024-11-05
I did a micro benchmark run, results:  http://benchmarks.glados.intel.com/d/1pXX4hUSz/microbenchmarks?orgId=1&var-tag=ci%7Creorder_dpas_gen&var-bench=All&var-device=Intel%28R%29%20Data%20Center%20GPU%20Max%201550&var-compiler=triton&var-backend=All&var-baseline_backend=triton-ci-XPU%201550&var-target_backend=triton-ci-XPU%201550

I was expecting to "See" and improvement for gemm with operand B transposed but performance seems about the same as CI (without this PR): 

![image](https://github.com/user-attachments/assets/ad639737-e8f9-4bc7-a464-e7443909ddb0)