@alexbaden 2024-11-05
The degradation appears to be due to the number of loads doubling for the B matrix. Previously we made 4 2D block read transpose loads:
```
  call spir_func void @_Z51intel_sub_group_2d_block_read_transpose_32b_32r8x1cPU3AS1viiiDv2_iPj(i8 addrspace(1)* readonly %1, i32 10240, i32 4096, i32 10240, <2 x i32> %215, i32* %216) #0
```
now we make 8:
```
  call spir_func void @_Z51intel_sub_group_2d_block_read_transpose_32b_16r8x1cPU3AS1viiiDv2_iPj(i8 addrspace(1)* readonly %1, i32 10240, i32 4096, i32 10240, <2 x i32> %219, i32* %220) #0
```
b/c the loads are smaller - half the number of rows - we need to do twice as many. 

Does something change with a newer IGC version that mitigates this?