@etiotto 2024-11-06
> Does something change with a newer IGC version that mitigates this?

Yes, the rationale for this change is discussed here: https://github.com/intel/intel-xpu-backend-for-triton/pull/2628#discussion_r1830298864