@chengjunlu 2024-11-11
I uses two IGC driver version to run the benchmark because:
  1. Using rolling IGC agama 1032 for GEMM AxB.T because there is a functional issue in DEV IGC for running the GEMM AxB.T.
  2. Using DEV IGC because it includes the latest enhancement for flash attention.

For the GEMM benchmark, the large block IO size is 1.09x better than the small block IO size in geomean.
For the flash attention benchmark, the large block IO size is 0.80x worse than the small block IO size in geomean.

The main reason of the small block IO size getting better performance in flash attention is because of there is no register spilling comparing to the large block IO size.

Here is the detail performance results for small/large block IO size comparison:
GEMM AxB.T

  | B | M | K | N | Tflops (small block IO) | Triton (large block IO) | ratio
-- | -- | -- | -- | -- | -- | -- | --
0 | 1 | 1024 | 1024 | 1024 | 47.892141 | 53.340376 | 1.113760523
1 | 1 | 2048 | 2048 | 2048 | 81.4289 | 92.454361 | 1.135399852
2 | 1 | 4096 | 4096 | 4096 | 106.56165 | 121.938171 | 1.144296949
3 | 1 | 8192 | 8192 | 8192 | 104.713196 | 116.605608 | 1.113571283
4 | 1 | 1 | 5120 | 13824 | 0.353559 | 0.351434 | 0.993989688
5 | 1 | 4 | 4096 | 12288 | 1.265171 | 1.264138 | 0.99918351
6 | 1 | 512 | 8192 | 8192 | 67.922074 | 74.869233 | 1.102281314
7 | 1 | 512 | 8192 | 32768 | 88.565153 | 99.414784 | 1.122504514
8 | 1 | 512 | 32768 | 8192 | 61.549843 | 69.834026 | 1.134593081
9 | 1 | 1024 | 16384 | 8192 | 71.986588 | 82.35006 | 1.143963928
10 | 1 | 1024 | 28672 | 8192 | 70.379633 | 71.105156 | 1.010308707
11 | 1 | 3072 | 4096 | 3072 | 96.256542 | 109.823859 | 1.140949557
12 | 1 | 4096 | 16384 | 8192 | 83.23504 | 98.109532 | 1.17870469
13 | 1 | 8192 | 16384 | 1024 | 75.496408 | 90.432862 | 1.19784324
14 | 1 | 8192 | 16384 | 4096 | 88.248834 | 98.666475 | 1.118048483
15 | 1 | 16384 | 1024 | 8192 | 122.578733 | 127.232374 | 1.037964506
16 | 1 | 16384 | 4096 | 8192 | 116.408825 | 122.843592 | 1.055277313
17 | 1 | 16384 | 8192 | 1024 | 105.810989 | 120.147341 | 1.135490199
18 | 1 | 16384 | 8192 | 4096 | 110.823567 | 123.97636 | 1.118682275
19 | 4 | 32768 | 128 | 4096 | 50.258884 | 50.281316 | 1.000446329
20 | 4 | 32768 | 4096 | 128 | 52.265309 | 61.017267 | 1.167452526
21 | 32 | 4096 | 4096 | 128 | 50.609409 | 53.296938 | 1.053103347
22 | 4096 | 8 | 128 | 16384 | 4.26713 | 4.285785 | 1.004371791
23 | 4096 | 8 | 16384 | 128 | 4.34589 | 4.354827 | 1.002056426

Flash Attention:

  | Z | H | N_CTX | D_HEAD | CAUSAL | Tflops (small block IO) | Triton (large block IO) | ratio
-- | -- | -- | -- | -- | -- | -- | -- | --
0 | 1 | 32 | 16384 | 64 | FALSE | 78.212688 | 61.394674 | 0.784970771
1 | 2 | 32 | 8192 | 64 | FALSE | 76.707298 | 59.493067 | 0.775585486
2 | 4 | 32 | 4096 | 64 | FALSE | 72.957003 | 57.576184 | 0.789179676
3 | 8 | 32 | 2048 | 64 | FALSE | 67.13312 | 54.010944 | 0.80453499
4 | 16 | 32 | 1024 | 64 | FALSE | 60.617356 | 50.008354 | 0.824984085
5 | 32 | 32 | 512 | 64 | FALSE | 62.716272 | 51.821513 | 0.826284971
6 | 4 | 48 | 1024 | 64 | FALSE | 70.047579 | 58.562413 | 0.836037645