@etiotto 2024-11-12
Discussed offline, decided to merge this PR with the env. variable so that we can more easily run performance experiments.